refactor table loader
