--- docs/docs/auto-docs/components/TableLoader/TableLoader/functions/default.md
@@ -1,19 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Function: default()
-
-> **default**(`props`): `Element`
-
-Defined in: [src/components/TableLoader/TableLoader.tsx:42](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/TableLoader/TableLoader.tsx#L42)
-
-## Parameters
-
-### props
-
-[`InterfaceTableLoader`](../interfaces/InterfaceTableLoader.md)
-
-## Returns
-
-`Element`
