--- docs/docs/auto-docs/components/TableLoader/TableLoader/interfaces/InterfaceTableLoader.md
@@ -1,39 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: InterfaceTableLoader
-
-Defined in: [src/components/TableLoader/TableLoader.tsx:35](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/TableLoader/TableLoader.tsx#L35)
-
-## Properties
-
-### data-testid?
-
-> `optional` **data-testid**: `string`
-
-Defined in: [src/components/TableLoader/TableLoader.tsx:39](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/TableLoader/TableLoader.tsx#L39)
-
-***
-
-### headerTitles?
-
-> `optional` **headerTitles**: `string`[]
-
-Defined in: [src/components/TableLoader/TableLoader.tsx:37](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/TableLoader/TableLoader.tsx#L37)
-
-***
-
-### noOfCols?
-
-> `optional` **noOfCols**: `number`
-
-Defined in: [src/components/TableLoader/TableLoader.tsx:38](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/TableLoader/TableLoader.tsx#L38)
-
-***
-
-### noOfRows
-
-> **noOfRows**: `number`
-
-Defined in: [src/components/TableLoader/TableLoader.tsx:36](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/TableLoader/TableLoader.tsx#L36)
