+++ docs/docs/auto-docs/shared-components/TableLoader/TableLoader/functions/default.md
@@ -1,0 +1,19 @@
+[Admin Docs](/)
+
+***
+
+# Function: default()
+
+> **default**(`props`): `Element`
+
+Defined in: src/shared-components/TableLoader/TableLoader.tsx:36
+
+## Parameters
+
+### props
+
+[`InterfaceTableLoader`](../../../../types/shared-components/TableLoader/interface/interfaces/InterfaceTableLoader.md)
+
+## Returns
+
+`Element`
