+++ docs/docs/auto-docs/types/shared-components/TableLoader/interface/interfaces/InterfaceTableLoader.md
@@ -1,0 +1,39 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceTableLoader
+
+Defined in: src/types/shared-components/TableLoader/interface.ts:1
+
+## Properties
+
+### data-testid?
+
+> `optional` **data-testid**: `string`
+
+Defined in: src/types/shared-components/TableLoader/interface.ts:5
+
+***
+
+### headerTitles?
+
+> `optional` **headerTitles**: `string`[]
+
+Defined in: src/types/shared-components/TableLoader/interface.ts:3
+
+***
+
+### noOfCols?
+
+> `optional` **noOfCols**: `number`
+
+Defined in: src/types/shared-components/TableLoader/interface.ts:4
+
+***
+
+### noOfRows
+
+> **noOfRows**: `number`
+
+Defined in: src/types/shared-components/TableLoader/interface.ts:2
