+++ src/screens/AdminPortal/OrganizationFunds/OrganizationFunds.tsx
@@ -11,7 +11,7 @@ import { useTranslation } from 'react-i18next';
 import { Navigate, useNavigate, useParams } from 'react-router';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import dayjs from 'dayjs';
-import TableLoader from 'components/TableLoader/TableLoader';
+import TableLoader from 'shared-components/TableLoader/TableLoader';
 import ReportingTable from 'shared-components/ReportingTable/ReportingTable';
 import FundModal from './modal/FundModal';
 import { FUND_LIST } from 'GraphQl/Queries/fundQueries';
@@ -385,9 +385,9 @@ const organizationFunds = (): JSX.Element => {
       </div>
 
       {!fundLoading &&
-      fundData &&
-      filteredAndSortedFunds.length === 0 &&
-      searchText.length > 0 ? (
+        fundData &&
+        filteredAndSortedFunds.length === 0 &&
+        searchText.length > 0 ? (
         <EmptyState
           icon={<Search />}
           message="noResultsFound"
