+++ src/screens/AdminPortal/Requests/Requests.tsx
@@ -13,7 +13,7 @@
  * @requires react-router-dom
  * @requires @mui/material
  * @requires GraphQl/Queries/Queries
- * @requires components/TableLoader/TableLoader
+ * @requires shared-components/TableLoader/TableLoader
  * @requires components/RequestsTableItem/RequestsTableItem
  * @requires subComponents/SearchBar
  * @requires utils/interfaces
@@ -61,7 +61,7 @@ import {
   MEMBERSHIP_REQUEST_PG,
   ORGANIZATION_LIST,
 } from 'GraphQl/Queries/Queries';
-import TableLoader from 'components/TableLoader/TableLoader';
+import TableLoader from 'shared-components/TableLoader/TableLoader';
 import { GridCellParams } from 'shared-components/DataGridWrapper';
 import type {
   ReportingTableColumn,
