+++ src/screens/AdminPortal/Users/Users.tsx
@@ -70,7 +70,7 @@ import {
   ORGANIZATION_LIST,
   USER_LIST_FOR_ADMIN,
 } from 'GraphQl/Queries/Queries';
-import TableLoader from 'components/TableLoader/TableLoader';
+import TableLoader from 'shared-components/TableLoader/TableLoader';
 import UsersTableItem from 'components/UsersTableItem/UsersTableItem';
 import InfiniteScroll from 'react-infinite-scroll-component';
 import type { InterfaceQueryUserListItem } from 'utils/interfaces';
