+++ src/shared-components/LoadingState/LoadingState.tsx
@@ -34,7 +34,7 @@ import { Spinner } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
 import styles from 'style/app-fixed.module.css';
 import type { InterfaceLoadingStateProps } from 'types/shared-components/LoadingState/interface';
-import TableLoader from 'components/TableLoader/TableLoader';
+import TableLoader from 'shared-components/TableLoader/TableLoader';
 
 const LoadingState = ({
   isLoading,
