+++ src/shared-components/TableLoader/TableLoader.module.css
