+++ src/types/shared-components/TableLoader/interface.ts
@@ -1,0 +1,6 @@
+export interface InterfaceTableLoader {
+    noOfRows: number;
+    headerTitles?: string[];
+    noOfCols?: number;
+    'data-testid'?: string;
+}
